pyproject.toml
[project]
name = "aperitif"
version = "0.1.0"
description = "Multiagent workflow for address extraction and color detection"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno>=0.1.0",
    "openai>=1.82.1",
    "pillow>=10.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
]
